requirements.txt
# Core Dependencies
streamlit>=1.28.0
torch>=2.0.0
transformers>=4.30.0
openai>=1.0.0

# Data Processing
python-dotenv>=1.0.0

# Database (Optional)
supabase>=2.0.0

# PDF Generation
reportlab>=4.0.0
markdown>=3.5.0

# Utilities
scikit-learn>=1.3.0
